--- src/TrimDB.Core/Storage/BlockCache/FileIdentifier.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace TrimDB.Core.Storage.BlockCache
-{
-    public struct FileIdentifier
-    {
-        public int Level;
-        public int FileId;
-    }
-}
